requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
mypy==1.7.1
django-stubs==4.2.7
types-redis==4.6.0.20241004
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
pytest-celery==0.0.0
factory-boy==3.3.0
freezegun==1.2.2 